Add DOM and SVG attributes to mapping of Preact style prop names
