+++ build-system/common/dom-svg-attributes.js
@@ -1,0 +1,108 @@
+/**
+ * @param {Object} obj
+ * @return {Object}
+ */
+function objectFlip(obj) {
+  const ret = {};
+  Object.keys(obj).forEach((key) => {
+    ret[obj[key]] = key;
+  });
+  return ret;
+}
+
+const DOM_ATTRIBUTES = {
+  className: 'class',
+  // TODO(wg-bento): Revert tabIndex with tabindex
+  tabindex: 'tabIndex',
+};
+
+const DOM_ATTRIBUTES_FLIPPED = objectFlip(DOM_ATTRIBUTES);
+
+/**
+ * This list derives all hyphenated attributes from
+ * https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute
+ */
+const SVG_ATTRIBUTES = {
+  accentHeight: 'accent-height',
+  alignmentBaseline: 'alignment-baseline',
+  arabicForm: 'arabic-form',
+  baselineShift: 'baseline-shift',
+  capHeight: 'cap-height',
+  clipPath: 'clip-path',
+  clipRule: 'clip-rule',
+  colorInterpolation: 'color-interpolation',
+  colorInterpolationFilters: 'color-interpolation-filters',
+  colorProfile: 'color-profile',
+  colorRendering: 'color-rendering',
+  dominantBaseline: 'dominant-baseline',
+  enableBackground: 'enable-background',
+  fillOpacity: 'fill-opacity',
+  fillRule: 'fill-rule',
+  floodColor: 'flood-color',
+  floodOpacity: 'flood-opacity',
+  fontFamily: 'font-family',
+  fontSize: 'font-size',
+  fontSizeAdjust: 'font-size-adjust',
+  fontStretch: 'font-stretch',
+  fontStyle: 'font-style',
+  fontVariant: 'font-variant',
+  fontWeight: 'font-weight',
+  glyphName: 'glyph-name',
+  glyphOrientationHorizontal: 'glyph-orientation-horizontal',
+  glyphOrientationVertical: 'glyph-orientation-vertical',
+  horizAdvX: 'horiz-adv-x',
+  horizOriginX: 'horiz-origin-x',
+  imageRendering: 'image-rendering',
+  letterSpacing: 'letter-spacing',
+  lightingColor: 'lighting-color',
+  markerEnd: 'marker-end',
+  markerMid: 'marker-mid',
+  markerStart: 'marker-start',
+  overlinePosition: 'overline-position',
+  overlineThickness: 'overline-thickness',
+  paintOrder: 'paint-order',
+  panose1: 'panose-1',
+  pointerEvents: 'pointer-events',
+  renderingIntent: 'rendering-intent',
+  shapeRendering: 'shape-rendering',
+  stopColor: 'stop-color',
+  stopOpacity: 'stop-opacity',
+  strikethroughPosition: 'strikethrough-position',
+  strikethroughThickness: 'strikethrough-thickness',
+  strokeDasharray: 'stroke-dasharray',
+  strokeDashoffset: 'stroke-dashoffset',
+  strokeLinecap: 'stroke-linecap',
+  strokeLinejoin: 'stroke-linejoin',
+  strokeMiterlimit: 'stroke-miterlimit',
+  strokeOpacity: 'stroke-opacity',
+  strokeWidth: 'stroke-width',
+  textAnchor: 'text-anchor',
+  textDecoration: 'text-decoration',
+  textRendering: 'text-rendering',
+  transformOrigin: 'transform-origin',
+  underlinePosition: 'underline-position',
+  underlineThickness: 'underline-thickness',
+  unicodeBidi: 'unicode-bidi',
+  unicodeRange: 'unicode-range',
+  unitsPerEm: 'units-per-em',
+  vAlphabetic: 'v-alphabetic',
+  vHanging: 'v-hanging',
+  vIdeographic: 'v-ideographic',
+  vMathematical: 'v-mathematical',
+  vectorEffect: 'vector-effect',
+  vertAdvY: 'vert-adv-y',
+  vertOriginX: 'vert-origin-x',
+  vertOriginY: 'vert-origin-y',
+  wordSpacing: 'word-spacing',
+  writingMode: 'writing-mode',
+  xHeight: 'x-height',
+};
+
+const SVG_ATTRIBUTES_FLIPPED = objectFlip(SVG_ATTRIBUTES);
+
+module.exports = {
+  DOM_ATTRIBUTES,
+  DOM_ATTRIBUTES_FLIPPED,
+  SVG_ATTRIBUTES,
+  SVG_ATTRIBUTES_FLIPPED,
+};
